refactor: Move SettingRow and add coming soon toast (#183)
- Move the `SettingRow` composable from `presentation/mypage/setting/components/SettingRow.kt` into `SettingScreen.kt` as a private function.
- Add a `ToastComingSoon` side effect to display a "feature coming soon" message.
- Implement click handlers for "Register Account," "Keyword Alert Setting," and "Safety Guide" to trigger the coming soon toast.
- Add `showComingSoonMessage` to the ViewModel to emit the side effect.
