+++ app/src/main/java/com/example/rentit/presentation/mypage/setting/SettingScreen.kt
@@ -1,23 +1,29 @@
 package com.example.rentit.presentation.mypage.setting
 
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Scaffold
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.semantics.contentDescription
+import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import com.example.rentit.R
 import com.example.rentit.common.component.CommonDivider
 import com.example.rentit.common.component.CommonTopAppBar
 import com.example.rentit.common.component.screenHorizontalPadding
 import com.example.rentit.common.theme.Gray300
 import com.example.rentit.common.theme.RentItTheme
-import com.example.rentit.presentation.mypage.setting.components.SettingRow
 
 @Composable
 fun SettingScreen(
@@ -46,6 +52,36 @@ fun SettingScreen(
 
             LogoutListItem(onLogoutClick)
         }
+    }
+}
+
+/**
+ * 설정 화면 List Item용 공통 Row
+ * (클릭, contentDescription 처리 포함)
+ */
+
+@Composable
+fun SettingRow(
+    label: String,
+    contentDesc: String = label,
+    isClickEnabled: Boolean = true,
+    onClick: () -> Unit = {},
+    endContent: @Composable (() -> Unit)? = null
+) {
+    Row(
+        modifier = Modifier
+            .semantics { contentDescription = contentDesc }
+            .height(65.dp)
+            .clickable(enabled = isClickEnabled) { onClick() }
+            .screenHorizontalPadding(),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Text(
+            modifier = Modifier.weight(1f),
+            text = label,
+            style = MaterialTheme.typography.bodyLarge
+        )
+        endContent?.invoke()
     }
 }
