+++ app/src/main/java/com/example/rentit/presentation/mypage/setting/SettingSideEffect.kt
@@ -3,4 +3,5 @@ package com.example.rentit.presentation.mypage.setting
 sealed class SettingSideEffect {
     data object NavigateToLogin: SettingSideEffect()
     data object ToastLogoutComplete: SettingSideEffect()
+    data object ToastComingSoon: SettingSideEffect()
 }
